[MT-2002] iOS: Remove CocoaPods artifacts and consolidate plugin classes
- Remove stale Pods/Pods.xcodeproj ref from xcworkspacedata in all example apps
- Remove empty PrivacyInfo.xcprivacy from tealium_moments_api and its podspec/Package.swift refs
- Rename SwiftTealiumPlugin -> TealiumPlugin (in-place), delete thin ObjC-bridging wrapper TealiumPlugin.swift
- Merge SwiftTealiumAdobeVisitorPlugin into TealiumAdobeVisitorPlugin, delete old file
- Merge SwiftTealiumFirebasePlugin into TealiumFirebasePlugin, delete old file
- Update all cross-module references (MomentsAPI, Braze, Firebase, VisitorDelegate, RunnerTests)
