+++ modules/tealium_adobevisitor/ios/tealium_adobevisitor/Sources/tealium_adobevisitor/TealiumFlutterAdobeExtentions.swift
@@ -11,10 +11,9 @@ import TealiumAdobeVisitorAPI
 extension AdobeVisitor {
     func asDictionary() -> [String: Any]? {
         guard let idSyncTTL = self.idSyncTTL, let idSyncTTL = Int(idSyncTTL),
-            let dcsRegion = self.dcsRegion, let dcsRegion = Int(dcsRegion),
-            let blob = self.blob,
-            let nextRefresh = self.nextRefresh,
-            let nextRefresh = Int(nextRefresh.unixTimeMilliseconds)
+              let dcsRegion = self.dcsRegion, let dcsRegion = Int(dcsRegion),
+              let blob = self.blob,
+              let nextRefresh = self.nextRefresh, let nextRefresh = Int(nextRefresh.unixTimeMilliseconds)
         else {
             return nil
         }
