+++ modules/tealium_moments_api/ios/tealium_moments_api/Package.swift
@@ -17,7 +17,7 @@ let package = Package(
         // tool rewrites this relative path to the versioned symlink at build time, so
         // it resolves both in this monorepo and when consumed from pub.dev.
         .package(name: "tealium", path: "../tealium"),
-        .package(url: "https://github.com/tealium/tealium-swift", from: "2.18.0")
+        .package(url: "https://github.com/tealium/tealium-swift", .upToNextMajor(from: "2.18.0"))
     ],
     targets: [
         .target(
@@ -28,9 +28,7 @@ let package = Package(
                 .product(name: "TealiumCore", package: "tealium-swift"),
                 .product(name: "TealiumMomentsAPI", package: "tealium-swift")
             ],
-            resources: [
-                .process("PrivacyInfo.xcprivacy")
-            ]
+
         )
     ]
 )
