+++ remotecommands/tealium_braze/ios/tealium_braze/Sources/tealium_braze/TealiumBrazePlugin.swift
@@ -11,7 +11,7 @@ import tealium
 
 public class TealiumBrazePlugin: NSObject, FlutterPlugin {
     public static func register(with registrar: FlutterPluginRegistrar) {
-        SwiftTealiumPlugin.registerRemoteCommandFactory(BrazeRemoteCommandFactory())
+        TealiumPlugin.registerRemoteCommandFactory(BrazeRemoteCommandFactory())
     }
 
     public func handle(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
