+++ remotecommands/tealium_firebase/ios/tealium_firebase/Sources/tealium_firebase/TealiumFirebasePlugin.swift
@@ -1,8 +1,14 @@
 import Flutter
+import UIKit
+import tealium
 
 @objc(TealiumFirebasePlugin)
 public class TealiumFirebasePlugin: NSObject, FlutterPlugin {
     public static func register(with registrar: FlutterPluginRegistrar) {
-        SwiftTealiumFirebasePlugin.register(with: registrar)
+        TealiumPlugin.registerRemoteCommandFactory(FirebaseRemoteCommandFactory())
+    }
+
+    public func handle(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
+        result(FlutterMethodNotImplemented)
     }
 }
