--- tealium/ios/tealium/Sources/tealium/TealiumPlugin.swift
@@ -1,8 +1,0 @@
-import Flutter
-
-@objc(TealiumPlugin)
-public class TealiumPlugin: NSObject, FlutterPlugin {
-    public static func register(with registrar: FlutterPluginRegistrar) {
-        SwiftTealiumPlugin.register(with: registrar)
-    }
-}
